refactor ocs sharing
